refactor the schema files
