+++ pkg/monitoring/avro_schemas.go
